+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+* [#4](https://github.com/persona-id/terraform-provider-stablepairer/pull/4)
+  Fix ModifyPlan handling of unknown values.
+  ([@drcapulet])
+
 ## 1.1.0
 
 * [#3](https://github.com/persona-id/terraform-provider-stablepairer/pull/3)